\name{fastimp-methods}
\docType{methods}
\alias{fastimp}
\alias{fastimp-methods}
\alias{fastimp,bigcforest-method}
\title{Methods for Function \code{fastimp} in Package \pkg{bigrf}}
\description{
Calculates a fast (Gini) variable importance measure by adding up the Gini decreases for each variable over all trees. The results are often consistent with the full variable importance calculated with \code{\link[bigrf]{varimp}}.
}
\usage{
  \S4method{fastimp}{bigcforest}(forest)
}
\arguments{
  \item{forest}{
    A random forest of class \code{"\linkS4class{bigcforest}"}.
  }
}
\section{Methods}{
\describe{

\item{\code{signature(forest = "bigcforest")}}{
Calculates the fast (Gini) variable importance for a classification random forest generated by \code{\link[bigrf]{bigrfc}}.
}
}}
\value{
A numeric vector containing the fast (Gini) variable importance measures for each variable.
}
\seealso{
\code{\link[bigrf]{varimp}}
}
\examples{
# Classify cars in the Cars93 data set by type (Compact, Large,
# Midsize, Small, Sporty, or Van).

# Load data.
data(Cars93, package="MASS")
x <- Cars93
y <- Cars93$Type

# Select variables with which to train model.
vars <- c(4:22)

# Run model, build 50 trees.
forest <- bigrfc(x, y, ntree=50L, varselect=vars)

# Calculate variable importance, including those for each out-of-bag example.
fastimportance <- fastimp(forest)
}
\keyword{methods}
\keyword{methods}
