\name{scaling-methods}
\docType{methods}
\alias{scaling}
\alias{scaling-methods}
\alias{scaling,bigrfprox-method}
\title{Methods for Function \code{scaling} in Package \pkg{bigrf}}
\description{
Calculates scaling coordinates for a random forest, based on the proximity matrix generated by \code{\link[bigrf]{proximities}}.
}
\usage{
\S4method{scaling}{bigrfprox}(prox, nscale=2L, trace=0L)
}
\arguments{
  \item{prox}{
    A proximity matrix of class \code{"\linkS4class{bigrfprox}"}.
  }
  \item{nscale}{The number of scaling co-ordinates to compute. Typically, the first two or three scaling co-ordinates are the most useful. Default: \code{2L}.}
  \item{trace}{
    \code{0} for no verbose output. \code{1} to print verbose output. Default: \code{0}.
  }
}
\value{
  A \code{matrix} containing the scaling co-ordinates for each example, where the \code{i}th column contains the \code{i}th scaling co-ordinates.
}
\section{Methods}{
\describe{

\item{\code{signature(prox = "bigrfprox")}}{
Calculates scaling coordinates for a random forest.
}
}}
\examples{
# Classify cars in the Cars93 data set by type (Compact, Large,
# Midsize, Small, Sporty, or Van).

# Load data.
data(Cars93, package="MASS")
x <- Cars93
y <- Cars93$Type

# Select variables with which to train model.
vars <- c(4:22)

# Run model, build 50 trees.
forest <- bigrfc(x, y, ntree=50L, varselect=vars)

# Calculate proximity matrix.
prox <- proximities(forest, y)

# Calculate scaling co-ordinates.
scale <- scaling(prox, 3L)

# Plot the 1st vs 2nd scaling co-ordinates.
plot(scale[, 1], scale[, 2], col=as.integer(y), pch=as.integer(y))

# Plot the 1st vs 3rd scaling co-ordinates.
plot(scale[, 1], scale[, 3], col=as.integer(y), pch=as.integer(y))

# Plot the 2nd vs 3rd scaling co-ordinates.
plot(scale[, 2], scale[, 3], col=as.integer(y), pch=as.integer(y))
}
\keyword{methods}
