\name{bigcforest-class}
\Rdversion{1.1}
\docType{class}
\alias{bigcforest-class}
\alias{show,bigcforest-method}
\alias{summary,bigcforest-method}

\title{Class \code{"bigcforest"}}
\description{
A random classification forest generated by \code{\link[bigrf]{bigrfc}}.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("bigcforest", ...)}, but most often is the result of a call to \code{\link[bigrf]{bigrfc}}.
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{"list"}. Each element is a \code{"\linkS4class{bigctree}"}, representing a tree in the random classification forest.}
    \item{\code{nexamples}:}{Object of class \code{"integer"}. Number of examples in the training set (including synthesized examples for unsupervised learning).}
    \item{\code{varselect}:}{Object of class \code{"integer"}. Indices of the columns of \code{x} that were used to train the model.}
    \item{\code{factorvars}:}{Object of class \code{"logical"}. Indicates which variables are factors or categorical (\code{TRUE})), and which are numeric (\code{FALSE}).}
    \item{\code{varnlevels}:}{Object of class \code{"integer"}. Number of levels in each categorical variable, or \code{0} for numeric variables.}
    \item{\code{contvarseq}:}{Object of class \code{"integer"}. Maps the continuous variables in \code{varselect} to the columns in \code{big.matrix} \code{a}. Meant for internal use by \code{bigrfc} or \code{grow} when growing trees.}
    \item{\code{ylevels}:}{Object of class \code{"character"}. Class labels.}
    \item{\code{ynclass}:}{Object of class \code{"integer"}. Number of classes.}
    \item{\code{ytable}:}{Object of class \code{"table"}. Counts of training examples in each class.}
    \item{\code{yclasswts}:}{Object of class \code{"matrix"}. One-dimensional matrix of scaled weights for each class.}
    \item{\code{ntrees}:}{Object of class \code{"integer"}. Number of trees in the forest.}
    \item{\code{maxnodes}:}{Object of class \code{"integer"}. Maximum number of nodes needed to grow trees fully. The actual number of nodes in each tree will be smaller than or equal to this.}
    \item{\code{nsplitvar}:}{Object of class \code{"integer"}. Number of variables to split on at each node.}
    \item{\code{maxndsize}:}{Object of class \code{"integer"}. Maximum number of examples in each node when growing the trees.}
    \item{\code{maxeslevels}:}{Object of class \code{"integer"}. Maximum number of levels for categorical variables for which exhaustive search of possible splits will be performed.}
    \item{\code{nrandsplit}:}{Object of class \code{"integer"} Number of random splits to examine for categorical variables with more than \code{maxeslevels} levels.}
    \item{\code{oobtimes}:}{Object of class \code{"integer"}. Number of times each training example has been out-of-bag.}
    \item{\code{oobvotes}:}{Object of class \code{"matrix"}. Out-of-bag votes for each training example.}
    \item{\code{oobpred}:}{Object of class \code{"integer"}. Out-of-bag predictions for each training example.}
    \item{\code{trainclserr}:}{Object of class \code{"numeric"}. Training errors of out-of-bag examples, by class.}
    \item{\code{trainerr}:}{Object of class \code{"numeric"}. Total training error of out-of-bag examples.}
    \item{\code{trainconfusion}:}{Object of class \code{"table"}. Confusion matrix for out-of-bag examples.}
    \item{\code{varginidec}:}{Object of class \code{"numeric"}. Decrease in Gini impurity for each variable over all trees.}
    \item{\code{cachepath}:}{Object of class \code{"character.or.NULL"}. Path to folder where data caches used in building the forest were stored, or \code{NULL} if data was processed completely in memory.}
  }
}
\section{Extends}{
Class \code{"\linkS4class{list}"}, from data part.
Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
}
\section{Methods}{
  \describe{
    \item{grow}{\code{signature(forest = "bigcforest")}: Grows more trees in the random forest. See \code{\link[bigrf]{grow}} for details.}
    \item{merge}{\code{signature(x = "bigcforest", y = "bigcforest")}: Merges two random forests into one. See \code{\link[bigrf]{merge}} for details.}
    \item{predict}{\code{signature(object = "bigcforest")}: Predicts the classes of a new test set using the random forest. See \code{\link[bigrf]{predict}} for details.}
    \item{varimp}{\code{signature(forest = "bigcforest")}: Calculates variable importance. See \code{\link[bigrf]{varimp}} for details.}
    \item{fastimp}{\code{signature(forest = "bigcforest")}: Calculates variable importance using a quicker but less accurate method. See \code{\link[bigrf]{fastimp}} for details.}
    \item{interactions}{\code{signature(forest = "bigcforest")}: Calculates variable interactions. See \code{\link[bigrf]{interactions}} for details.}
    \item{proximities}{\code{signature(forest = "bigcforest")}: Calculates the proximity matrix for the random forest. See \code{\link[bigrf]{proximities}} for details.}
    \item{show}{\code{signature(object = "bigcforest")}: Prints the random forest.}
    \item{summary}{\code{signature(object = "bigcforest")}: Prints summary information on the random forest, including out-of-bag training error estimates and the confusion matrix.}
	 }
}
\examples{
showClass("bigcforest")
}
\keyword{classes}
\keyword{models}
\keyword{tree}
